Broke out all the events into their own files within an events folder for better organization going forwards.
