+++ src/events/mod.rs
@@ -1,0 +1,10 @@
+mod runner;
+mod normal_mode;
+mod add_edit_mode;
+mod filter_mode;
+mod summary_mode;
+mod selection_mode;
+mod settings_mode;
+mod recurring_mode;
+
+pub use runner::run_app;
