+++ src/main.rs
@@ -1,6 +1,6 @@
 mod app;
 mod config;
-mod event;
+mod events;
 mod model;
 mod persistence;
 mod recurring;
@@ -8,7 +8,7 @@ mod ui;
 mod validation;
 
 use crate::app::state::App;
-use event::run_app;
+use events::run_app;
 use persistence::save_transactions;
 
 use crossterm::{
